Reorganiza o application controller e reduz o uso do sleep nos testes.
